requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.3
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-swagger==2.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
httpie==3.2.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mysqlclient==2.1.1
openapi-codec==1.3.2
openpyxl==3.0.10
psycopg2==2.9.5
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
PyMySQL==1.0.2
PySocks==1.7.1
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
simplejson==3.18.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
